- Fix lightsans Rust compatible

+++ etudelib/models/lightsans/torch_model.py
@@ -100,6 +100,6 @@ class LightSANsModel(nn.Module):
 
     def gather_indexes(self, output, gather_index):
         """Gathers the vectors at the specific positions over a minibatch"""
-        gather_index = gather_index.view(-1, 1, 1).expand(-1, -1, output.shape[-1])
+        gather_index = gather_index.view(-1, 1, 1).expand(-1, -1, output.shape[-1]).to(torch.int64)
         output_tensor = output.gather(dim=1, index=gather_index)
         return output_tensor.squeeze(1)
